refactor(clerk-js): Move CurrentPlanSection and PaymentMethodSection to common
