+++ packages/clerk-js/src/ui/common/index.ts
@@ -16,3 +16,5 @@ export * from './PrintableComponent';
 export * from './NotificationCountBadge';
 export * from './RemoveResourceForm';
 export * from './QRCode';
+export * from './CurrentPlanSection';
+export * from './PaymentMethodSection';
